+++ _src/visitor_test.go
