+++ visitor.go
@@ -1,87 +1,30 @@
-package config
+package configkit
 
-import (
-	"context"
-)
+import "context"
 
-// Visitor is an interface for walking application configurations.
+// Visitor is a visitor that visits configurations.
 type Visitor interface {
-	// VisitApplicationConfig is called by ApplicationConfig.Accept().
-	VisitApplicationConfig(context.Context, *ApplicationConfig) error
-
-	// VisitAggregateConfig is called by AggregateConfig.Accept().
-	VisitAggregateConfig(context.Context, *AggregateConfig) error
-
-	// VisitProcessConfig is called by ProcessConfig.Accept().
-	VisitProcessConfig(context.Context, *ProcessConfig) error
-
-	// VisitIntegrationConfig is called by IntegrationConfig.Accept().
-	VisitIntegrationConfig(context.Context, *IntegrationConfig) error
-
-	// VisitProjectionConfig is called by ProjectionConfig.Accept().
-	VisitProjectionConfig(context.Context, *ProjectionConfig) error
+	VisitApplication(context.Context, Application) error
+	VisitAggregate(context.Context, Aggregate) error
+	VisitProcess(context.Context, Process) error
+	VisitIntegration(context.Context, Integration) error
+	VisitProjection(context.Context, Projection) error
 }
 
-// FuncVisitor is an implementation of Visitor that dispatches to regular
-// functions.
-type FuncVisitor struct {
-	// ApplicationConfig, if non-nil, is called by VisitApplicationConfig().
-	ApplicationConfig func(context.Context, *ApplicationConfig) error
-
-	// AggregateConfig, if non-nil, is called by VisitAggregateConfig().
-	AggregateConfig func(context.Context, *AggregateConfig) error
-
-	// ProcessConfig, if non-nil, is called by VisitProcessConfig().
-	ProcessConfig func(context.Context, *ProcessConfig) error
-
-	// IntegrationConfig, if non-nil, is called by VisitIntegrationConfig().
-	IntegrationConfig func(context.Context, *IntegrationConfig) error
-
-	// ProjectionConfig, if non-nil, is called by VisitProjectionConfig().
-	ProjectionConfig func(context.Context, *ProjectionConfig) error
+// PortableVisitor is a visitor that visits "portable" configurations.
+type PortableVisitor interface {
+	VisitPortableApplication(context.Context, PortableApplication) error
+	VisitPortableAggregate(context.Context, PortableAggregate) error
+	VisitPortableProcess(context.Context, PortableProcess) error
+	VisitPortableIntegration(context.Context, PortableIntegration) error
+	VisitPortableProjection(context.Context, PortableProjection) error
 }
 
-// VisitApplicationConfig calls v.ApplicationConfig if it is non-nil.
-func (v *FuncVisitor) VisitApplicationConfig(ctx context.Context, cfg *ApplicationConfig) error {
-	if v.ApplicationConfig != nil {
-		return v.ApplicationConfig(ctx, cfg)
-	}
-
-	return nil
-}
-
-// VisitAggregateConfig calls v.AggregateConfig if it is non-nil.
-func (v *FuncVisitor) VisitAggregateConfig(ctx context.Context, cfg *AggregateConfig) error {
-	if v.AggregateConfig != nil {
-		return v.AggregateConfig(ctx, cfg)
-	}
-
-	return nil
-}
-
-// VisitProcessConfig calls v.ProcessConfig if it is non-nil.
-func (v *FuncVisitor) VisitProcessConfig(ctx context.Context, cfg *ProcessConfig) error {
-	if v.ProcessConfig != nil {
-		return v.ProcessConfig(ctx, cfg)
-	}
-
-	return nil
-}
-
-// VisitIntegrationConfig calls v.IntegrationConfig if it is non-nil.
-func (v *FuncVisitor) VisitIntegrationConfig(ctx context.Context, cfg *IntegrationConfig) error {
-	if v.IntegrationConfig != nil {
-		return v.IntegrationConfig(ctx, cfg)
-	}
-
-	return nil
-}
-
-// VisitProjectionConfig calls v.ProjectionConfig if it is non-nil.
-func (v *FuncVisitor) VisitProjectionConfig(ctx context.Context, cfg *ProjectionConfig) error {
-	if v.ProjectionConfig != nil {
-		return v.ProjectionConfig(ctx, cfg)
-	}
-
-	return nil
+// RichVisitor is a visitor that visits "rich" configurations.
+type RichVisitor interface {
+	VisitRichApplication(context.Context, RichApplication) error
+	VisitRichAggregate(context.Context, RichAggregate) error
+	VisitRichProcess(context.Context, RichProcess) error
+	VisitRichIntegration(context.Context, RichIntegration) error
+	VisitRichProjection(context.Context, RichProjection) error
 }
